Added -Wall -Wextra -Werror for linux and /WX /W2 for windows in aiebu lib and utility (#74)

+++ src/cpp/aiebu/src/assembler/aiebu_assembler.cpp
@@ -122,7 +122,7 @@ aiebu_assembler_get_elf(enum aiebu_assembler_buffer_type type,
     char *aelf = static_cast<char*>(std::malloc(sizeof(char)*velf.size()));
     std::copy(velf.begin(), velf.end(), aelf);
     *elf_buf = (void*)aelf;
-    ret = velf.size();
+    ret =  static_cast<int>(velf.size());
   }
   catch (aiebu::error &ex)
   {
